payments/requirements.txt
starlette==0.52.1
fastapi==0.128.2
uvicorn==0.40.0
redis-om==1.0.6
python-dotenv==1.2.1
requests==2.32.5
